Housekeeping / break out tags and instantiate functionality for reuse
