+++ lib/built.rb
@@ -52,6 +52,9 @@ require "built/constants"
 require "built/i18n"
 require "built/util"
 require "built/error"
+require "built/tags"
+require "built/instantiate"
+
 require "built/client"
 require "built/basic_object"
 require "built/application"
